Move mod action enum to enums module
